+++ src/time-zone/time-zone.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, Logger } from "@nestjs/common";
 import { ChatSettingName } from "@prisma/client";
 import { format, getTimezoneOffset } from "date-fns-tz";
 import { changeLanguage, t } from "i18next";
@@ -12,10 +12,13 @@ import { buildCbData, getChatHtmlLink, getPagination, parseCbData } from "src/ut
 import type { InlineKeyboardButton } from "telegraf/typings/core/types/typegram";
 
 import { TimeZoneAction } from "./interfaces/action.interface";
+import type { TimeZoneRenderSettingsOptions } from "./interfaces/settings-options.interface";
 
 @Update()
 @Injectable()
 export class TimeZoneService {
+  private readonly logger = new Logger(TimeZoneService.name);
+
   /**
    * Creates service
    * @param prismaService Database service
@@ -39,7 +42,7 @@ export class TimeZoneService {
         await this.saveSettings(ctx, chatId, value);
         break;
       case TimeZoneAction.SETTINGS:
-        await this.renderSettings(ctx, chatId, skip);
+        await this.renderSettings(ctx, { chatId, shouldAnswerCallback: true, skip });
         break;
       default:
         await next();
@@ -47,37 +50,30 @@ export class TimeZoneService {
   }
 
   /**
-   * Gets all available time zone identifiers
+   * Gets all available time zone identifiers sorted by offset
    * @returns Time zone identifier
    */
   private getAllTimeZones(): string[] {
     return Intl.supportedValuesOf("timeZone").sort((a, b) => {
       const aOffset = getTimezoneOffset(a);
       const bOffset = getTimezoneOffset(b);
-      // Sort by offset
       if (aOffset < bOffset) {
         return -1;
       }
-      if (aOffset > bOffset) {
-        return 1;
-      }
-      // Then sort by name
-      if (a < b) {
-        return -1;
-      }
-      return a > b ? 1 : 0;
+      return aOffset > bOffset ? 1 : 0;
     });
   }
 
   /**
    * Renders settings
    * @param ctx Callback context
-   * @param chatId Id of the chat which is edited
-   * @param skip Skip count
+   * @param options Render options
    */
-  private async renderSettings(ctx: CallbackCtx, chatId: number, skip?: number): Promise<void> {
+  private async renderSettings(ctx: CallbackCtx, options: TimeZoneRenderSettingsOptions): Promise<void> {
+    const { chatId, shouldAnswerCallback, skip } = options;
     if (!ctx.chat || isNaN(chatId)) {
-      throw new Error("Chat is not defined to render time zone settings.");
+      this.logger.error("Chat is not defined to render time zone settings");
+      return;
     }
 
     const { language } = await this.prismaService.upsertChatWithCache(ctx.chat, ctx.callbackQuery.from);
@@ -91,12 +87,12 @@ export class TimeZoneService {
     const timeZones = this.getAllTimeZones();
     const valueIndex = timeZones.indexOf(dbChat.timeZone);
     // Use provided skip or the index of the current value. Use 0 as the last fallback.
-    const patchedSkip = skip ?? (valueIndex > -1 ? valueIndex : 0);
+    const patchedSkip = skip ?? (valueIndex > -1 ? Math.floor(valueIndex / PAGE_SIZE) * PAGE_SIZE : 0);
     const value = `${format(Date.now(), "O", { timeZone: dbChat.timeZone })} ${dbChat.timeZone}`;
     const msg = t("timeZone:select", { CHAT: chatLink, VALUE: value });
 
     await Promise.all([
-      ctx.answerCbQuery(),
+      shouldAnswerCallback && ctx.answerCbQuery(),
       ctx.editMessageText(this.prismaService.joinModifiedInfo(msg, ChatSettingName.TIME_ZONE, dbChat), {
         parse_mode: "HTML",
         reply_markup: {
@@ -121,15 +117,6 @@ export class TimeZoneService {
   }
 
   /**
-   * Sanitizes adding bots rule
-   * @param value Value
-   * @returns Sanitized value
-   */
-  private sanitizeValue(value: string | null): string {
-    return value && Intl.supportedValuesOf("timeZone").includes(value) ? value : "Etc/UTC";
-  }
-
-  /**
    * Saves settings
    * @param ctx Callback context
    * @param chatId Id of the chat which is edited
@@ -137,7 +124,8 @@ export class TimeZoneService {
    */
   private async saveSettings(ctx: CallbackCtx, chatId: number, value: string | null): Promise<void> {
     if (!ctx.chat || isNaN(chatId)) {
-      throw new Error("Chat is not defined to save time zone settings.");
+      this.logger.error("Chat is not defined to save time zone settings");
+      return;
     }
 
     const { language } = await this.prismaService.upsertChatWithCache(ctx.chat, ctx.callbackQuery.from);
@@ -147,14 +135,13 @@ export class TimeZoneService {
       return; // The user is no longer an administrator, or the bot has been banned from the chat.
     }
 
-    const timeZone = this.sanitizeValue(value);
-    const skip = Math.max(0, Math.floor(this.getAllTimeZones().indexOf(timeZone) / PAGE_SIZE) * PAGE_SIZE);
+    const timeZone = value && Intl.supportedValuesOf("timeZone").includes(value) ? value : "Etc/UTC";
 
     await this.prismaService.$transaction([
       this.prismaService.chat.update({ data: { timeZone }, select: { id: true }, where: { id: chatId } }),
       this.prismaService.upsertChatSettingsHistory(chatId, ctx.callbackQuery.from.id, ChatSettingName.TIME_ZONE),
     ]);
     await this.prismaService.deleteChatCache(chatId);
-    await Promise.all([this.settingsService.notifyChangesSaved(ctx), this.renderSettings(ctx, chatId, skip)]);
+    await Promise.all([this.settingsService.notifyChangesSaved(ctx), this.renderSettings(ctx, { chatId })]);
   }
 }