+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.19.0 (2024-03-10)
+
+- Improvement: profanity filter now checks edited messages
+- Improvement: better formatting for voteban results
+- Fix: pagination in time zone settings
+- Other: internal improvements
+
 ## 1.18.0 (2024-02-29)
 
 - Improvement: if a message consists of several photos, now they will all be deleted using the /warn command